Add 'BackToGame' function to RespawnMenu
The commit introduced the 'BackToGame' function in the RespawnMenu class. It allows players to return to the game from the respawn menu. This method removes the current level from the world and resumes the game by changing the pause state of the player controller. Additionally, the RespawnMenu class is now defined in the SideScrollerGameInstance.