Update level switching logic in LevelGameMode
This commit includes changes to the level switching logic in LevelGameMode.cpp. The code now uses actual level numbers instead of checking the existence of the level file. As a result, unnecessary file I/O is eliminated, which should improve the performance of level switching. Additionally, the error message when all levels are completed has been simplified.